Refactor of dbModel functions - issue #1119
