+++ src/lib/orionld/dbModel/dbModelFromApiKeyValues.cpp
@@ -1,0 +1,58 @@
+/*
+*
+* Copyright 2022 FIWARE Foundation e.V.
+*
+* This file is part of Orion-LD Context Broker.
+*
+* Orion-LD Context Broker is free software: you can redistribute it and/or
+* modify it under the terms of the GNU Affero General Public License as
+* published by the Free Software Foundation, either version 3 of the
+* License, or (at your option) any later version.
+*
+* Orion-LD Context Broker is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero
+* General Public License for more details.
+*
+* You should have received a copy of the GNU Affero General Public License
+* along with Orion-LD Context Broker. If not, see http://www.gnu.org/licenses/.
+*
+* For those usages not covered by this license please contact with
+* orionld at fiware dot org
+*
+* Author: Ken Zangelin
+*/
+extern "C"
+{
+#include "kjson/KjNode.h"                                      // KjNode
+#include "kjson/kjLookup.h"                                    // kjLookup
+#include "kjson/kjBuilder.h"                                   // kjObject, kjChildRemove, kjChildAdd,
+}
+
+#include "orionld/common/orionldState.h"                       // orionldState
+#include "orionld/dbModel/dbModelFromApiKeyValues.h"           // Own interface
+
+
+
+// -----------------------------------------------------------------------------
+//
+// dbModelFromApiKeyValues -
+//
+KjNode* dbModelFromApiKeyValues(KjNode* kvObjectArray, const char* name)
+{
+  KjNode* kvs = kjObject(orionldState.kjsonP, name);
+
+  for (KjNode* kv = kvObjectArray->value.firstChildP; kv != NULL; kv = kv->next)
+  {
+    KjNode* key   = kjLookup(kv, "key");
+    KjNode* value = kjLookup(kv, "value");
+
+    // Steal 'value' from 'kv'
+    kjChildRemove(kv, value);
+
+    value->name = key->value.s;
+    kjChildAdd(kvs, value);
+  }
+
+  return kvs;
+}
